requirements.txt
# Core dependencies
streamlit>=1.32.0
python-dotenv>=1.0.0

# LangChain and related packages
langchain>=0.0.350
langchain-community>=0.0.10
langchain-core>=0.0.10
langchain-huggingface>=0.0.5

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Hugging Face dependencies
huggingface-hub>=0.19.0
transformers>=4.36.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Additional utilities
requests>=2.31.0
tqdm>=4.65.0
pyyaml>=6.0
aiohttp>=3.8.0
urllib3>=2.0.0

# Optional dependencies for better performance
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0